Add /results endpoint
Factor common check validation functionality in client/utils.rb
and then include that functionality in both the raw socket.rb
and http_socket.rb components.
