+++ lib/sensu/client/socket.rb
@@ -1,4 +1,5 @@
 require "sensu/json"
+require "sensu/client/utils"
 
 module Sensu
   module Client
@@ -47,7 +48,7 @@ module Sensu
     # of data was received, the agent will give up on the sender, and
     # instead respond +"invalid"+ and close the connection.
     class Socket < EM::Connection
-      class DataError < StandardError; end
+      include CheckUtils
 
       attr_accessor :logger, :settings, :transport, :protocol
 
@@ -115,64 +116,6 @@ module Sensu
         end
       end
 
-      # Validate check result attributes.
-      #
-      # @param [Hash] check result to validate.
-      def validate_check_result(check)
-        unless check[:name] =~ /\A[\w\.-]+\z/
-          raise DataError, "check name must be a string and cannot contain spaces or special characters"
-        end
-        unless check[:source].nil? || check[:source] =~ /\A[\w\.-]+\z/
-          raise DataError, "check source must be a string and cannot contain spaces or special characters"
-        end
-        unless check[:output].is_a?(String)
-          raise DataError, "check output must be a string"
-        end
-        unless check[:status].is_a?(Integer)
-          raise DataError, "check status must be an integer"
-        end
-        unless check[:executed].is_a?(Integer)
-          raise DataError, "check executed timestamp must be an integer"
-        end
-        unless check[:ttl].nil? || (check[:ttl].is_a?(Integer) && check[:ttl] > 0)
-          raise DataError, "check ttl must be an integer greater than 0"
-        end
-      end
-
-      # Publish a check result to the Sensu transport.
-      #
-      # @param [Hash] check result.
-      def publish_check_result(check)
-        payload = {
-          :client => @settings[:client][:name],
-          :check => check.merge(:issued => Time.now.to_i)
-        }
-        payload[:signature] = @settings[:client][:signature] if @settings[:client][:signature]
-        @logger.info("publishing check result", :payload => payload)
-        @transport.publish(:direct, "results", Sensu::JSON.dump(payload)) do |info|
-          if info[:error]
-            @logger.error("failed to publish check result", {
-              :payload => payload,
-              :error => info[:error].to_s
-            })
-          end
-        end
-      end
-
-      # Process a check result. Set check result attribute defaults,
-      # validate the attributes, publish the check result to the Sensu
-      # transport, and respond to the sender with the message +"ok"+.
-      #
-      # @param [Hash] check result to be validated and published.
-      # @raise [DataError] if +check+ is invalid.
-      def process_check_result(check)
-        check[:status] ||= 0
-        check[:executed] ||= Time.now.to_i
-        validate_check_result(check)
-        publish_check_result(check)
-        respond("ok")
-      end
-
       # Parse a JSON check result. For UDP, immediately raise a parser
       # error. For TCP, record parser errors, so the connection
       # +watchdog+ can report them.
@@ -183,6 +126,7 @@ module Sensu
           check = Sensu::JSON.load(data)
           cancel_watchdog
           process_check_result(check)
+          respond("ok")
         rescue Sensu::JSON::ParseError, ArgumentError => error
           if @protocol == :tcp
             @parse_error = error.to_s
